Add very simple test mode that outputs screenshots for later comparison
Add a very simple test mode which outputs PNG screenshots of each frame to the current directory. This is useful to make sure that ep.c code changes/cleanup are not affecting the output.

Note that these screenshots are very dependent on GPU/driver/OS--in my experience they aren't even the same between 2 GPUs (integrated/discrete) on the same laptop!
Make sure to generate the reference/test PNGs with the same configuration.

- Compile using 'make test'
- electroportis_glut will run for 10000 frames and dump screenshots in the current directory
- Make changes and compile/run again (in a different directory)
- 'diff -r' the two directories to determine if the output changed