ep.c: display(): Simplify code incrementing 'n'
'n' is an int that starts at 0, goes up to 127, then wraps around to 0. The absolute value is being computed (through the shift/XOR/subtraction) after it's incremented and before it is ANDed with 127 but there's no need for this because it's always positive.
The right shift always results in 0 (because 'n' can only be between 0 and 127), so the XOR and subtraction have no effect on 'n'