ep.c: Fix buffer overrun in readAnimation()
Fix a buffer overflow in readAnimation(). At the end of reading the defaultScript, s0 points to the null-terminator, 's0++' increments it over the end of the buffer, and then strlen() starts reading from past the end of the string.
Instead, if s0 is already at the end of the string, jump rather than incrementing it.

On 64-bit, since the VAO commit, this was causing pretty crazy color effects (try it sometime).
Found with address sanitizer.